FUN-24 Move molecules components to "common" and "page-specific" folders specifically
