+++ client/components/Molecules/common/TheProductImagesGallery.vue
