+++ client/components/Molecules/common/TheRating.vue
