+++ client/components/Molecules/common/items/TheProductItem.vue
