+++ client/layouts/default.vue
@@ -22,7 +22,7 @@
 import TheMainHeader from '@/components/Organisms/base/TheMainHeader.vue'
 import TheMainContent from '@/components/Organisms/base/TheMainContent.vue'
 import TheMainFooter from '@/components/Organisms/base/TheMainFooter.vue'
-import ThePageHeader from '@/components/Molecules/ThePageHeader.vue'
+import ThePageHeader from '@/components/Molecules/common/ThePageHeader.vue'
 
 const breabcrumbs = [
   {
