+++ client/layouts/list.vue
@@ -25,7 +25,7 @@ import TheMainHeader from '@/components/Organisms/base/TheMainHeader.vue'
 import TheMainSidePanel from '@/components/Organisms/base/TheMainSidePanel.vue'
 import TheMainContent from '@/components/Organisms/base/TheMainContent.vue'
 import TheMainFooter from '@/components/Organisms/base/TheMainFooter.vue'
-import ThePageHeader from '@/components/Molecules/ThePageHeader.vue'
+import ThePageHeader from '@/components/Molecules/common/ThePageHeader.vue'
 
 const breabcrumbs = [
   {
